Update and rename MEASUREMENT.py to measurement.py
pylint measurement.py
--------------------------------------------------------------------
Your code has been rated at 10.00/10 (previous run: 10.00/10, +0.00)
